Added sdcard module using Petit FatFS
Integrated new sd card module into the bootloader and
tested it. Bootloader mode is currently controled by a
switch attached to pin RA1. In bootloader mode the firmware
will look for a file called platform.hex on the root of
the SD card and load that into the user program space.